+++ crates/cli/src/enums.rs
@@ -1,4 +1,5 @@
 use clap::ArgEnum;
+use serde::{Deserialize, Serialize};
 use strum_macros::Display;
 
 #[derive(ArgEnum, Clone, Debug, Default, Display)]
@@ -18,4 +19,17 @@ pub enum LogLevel {
     Info,
     Debug,
     Trace,
+}
+
+#[derive(ArgEnum, Clone, Copy, Debug, Deserialize, Display, Default, PartialEq, Serialize)]
+#[serde(rename_all = "lowercase")]
+pub enum TouchedStatus {
+    Added,
+    #[default]
+    All,
+    Deleted,
+    Modified,
+    Staged,
+    Unstaged,
+    Untracked,
 }