+++ packages/cli/CHANGELOG.md
@@ -9,7 +9,7 @@
   - To start, `moon query touched-files` can be used to query touched files. The same files
     `moon ci` and `moon run` use.
   - Also `moon query projects` can be used to query about projects in the project graph.
-- Added `bash` as a support value for the project `language` setting.
+- Added `bash` as a supported value for the project `language` setting.
 
 #### 🐞 Fixes
 
@@ -21,6 +21,8 @@
 
 - Updated Rust to v1.62.
 - Refactored our action runner to support additional languages in the future.
+- Refactored Windows to execute package binaries with `node.exe` directly, instead of with 
+  `cmd.exe` + the `.bin/*.cmd` file.
 
 ## 0.5.0